Update/consolidate aabb-collider, now box-collider
Based on MQTT pub behavior limited to client head and hands, the primary
collision tracking component is now called `box-collider`, and collidable
objects now should be designated with `box-collision-listener`. A `dynamic`
attribute of `box-collision-listener` should be set for objects that move
such that their bounding boxes need to be regularly recalculated.