Apply overrides to camObj, camFollow no controls
Overrides in non-XR/rigged situations imply that all camera controls are
relinquished, therefore no look-controls should be active to require a
respective pitch/yaw rig update.

handleCameraOverride still does check for rotation key so this could still
allow for omitting rotation in a new position-tracking-only override in nonXR
sessions, for example with a seat manager in which camera control is retained
by the user.