Stopgap fix for likely race condition w/ gUM
Previously armode implied noav=1, so gUM from jitsi connect
was not an issue. Currently it appears that both webAR gUM
and jitsi gUM may be overlapping, leading to a blank canvas
layer where the webcam texture is textured. As a stopgap,
we just induce artificial delay, which largely overlaps
with the series of popups asking for video permissions.

TODO: Figure out a way to properly emit and listen to events
in sane way for gUM-based passthrough and jitsi