fix: push-based mqtt message handler
This is the handler that was registered to handle updates when the
tab/window was inactive, so that mqtt messages could be still
processed in the background without RAF firing tocks. The was not refactored
correctly when batched messages/messageQueues were added, and still tried to
handle JSON-or-not message validation (now handled by worker onMsg dispatcher).
What this handler function actually should have been taking in was
an array of already json-validated messages.

This was apparently silently failing :(

This now properly fixes #668