Add submodel-parent component
When a component has a parent obj of a hierarchical model, sets the object3d
parent as the matching named subcomponent of that model, e.g. "joint_t_6".
This allows, for instance, a box to 'follow' the leg of an armature rather
than just the base pose of the model (probably the torso).